# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DK Option Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listDailyReports
      tags: [portfolio]
      summary: List persisted daily reports (newest first)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Daily report history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReportSummary"
    post:
      operationId: generateDailyReport
      tags: [portfolio]
      summary: Generate and persist a new daily report
      responses:
        "200":
          description: The generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
    delete:
      operationId: clearDailyReports
      tags: [portfolio]
      summary: Delete all persisted daily reports
      responses:
        "200":
          description: Reports cleared (returns the deleted rows so they can be restored)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearReportsResult"
  /reports/restore:
    post:
      operationId: restoreDailyReports
      tags: [portfolio]
      summary: Restore previously cleared daily reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreReportsRequest"
      responses:
        "200":
          description: Reports restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreReportsResult"
  /reports/compare-narration:
    post:
      operationId: narrateReportComparison
      tags: [portfolio]
      summary: AI-narrated plain-English summary of what changed between two reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareReportsRequest"
      responses:
        "200":
          description: Narrated comparison (deltas remain deterministic on the client)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareReportsResponse"
        "400":
          description: Invalid request
        "404":
          description: One or both reports not found
  /reports/{id}:
    get:
      operationId: getDailyReport
      tags: [portfolio]
      summary: Get a single persisted daily report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
    delete:
      operationId: deleteDailyReport
      tags: [portfolio]
      summary: Delete a persisted daily report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteReportResult"

  # ─── Value Investing (Buffett-style research & education) ──────────────────────
...
components:
  schemas:
...
    ScoreCard:
      type: object
      required: [score, label, detail]
      properties:
        score:
          type: integer
        label:
          type: string
        detail:
          type: string
    PortfolioHealthComponent:
      type: object
      required: [key, label, score, weight, detail]
      properties:
        key:
          type: string
        label:
          type: string
        score:
          type: integer
        weight:
          type: number
        detail:
          type: string
    ThreatCounts:
      type: object
      required: [green, yellow, red]
      properties:
        green:
          type: integer
        yellow:
          type: integer
        red:
          type: integer
    PortfolioHealth:
      type: object
      required:
        [health, exposure, riskConcentration, components, netDelta, netTheta, netVega, netGamma, largestSymbolPct, threatCounts]
      properties:
        health:
          $ref: "#/components/schemas/ScoreCard"
        exposure:
          $ref: "#/components/schemas/ScoreCard"
        riskConcentration:
          $ref: "#/components/schemas/ScoreCard"
        components:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioHealthComponent"
        netDelta:
          type: number
        netTheta:
          type: number
        netVega:
          type: number
        netGamma:
          type: number
        largestSymbol:
          type: string
          nullable: true
        largestSymbolPct:
          type: number
        threatCounts:
          $ref: "#/components/schemas/ThreatCounts"
    MarketBriefingEvent:
      type: object
      required: [type, label, date, daysAway, impact]
      properties:
        type:
          type: string
        label:
          type: string
        date:
          type: string
        daysAway:
          type: integer
        impact:
          type: string
          enum: [low, medium, high]
        symbol:
          type: string
          nullable: true
    MarketBriefing:
      type: object
      required:
        [date, regime, regimeLabel, syntheticVix, vixLabel, avgIvRank, breadth, ivEnvironment, headline, drivers, catalysts]
      properties:
        date:
          type: string
        regime:
          type: string
          enum: [risk_on, neutral, risk_off]
        regimeLabel:
          type: string
        syntheticVix:
          type: number
        vixLabel:
          type: string
        avgIvRank:
          type: integer
        breadth:
          type: integer
        ivEnvironment:
          type: string
        headline:
          type: string
        drivers:
          type: array
          items:
            type: string
        catalysts:
          type: array
          items:
            $ref: "#/components/schemas/MarketBriefingEvent"
...
    ReportSummary:
      type: object
      required:
        [accountValue, openPositions, totalUnrealizedPnl, netDelta, netTheta, netVega, netGamma, portfolioRiskUsedPct, maxPortfolioRiskPct, withinLimits, thetaDaily, thetaMonthly]
      properties:
        accountValue:
          type: number
        openPositions:
          type: integer
        totalUnrealizedPnl:
          type: number
        netDelta:
          type: number
        netTheta:
          type: number
        netVega:
          type: number
        netGamma:
          type: number
        portfolioRiskUsedPct:
          type: number
        maxPortfolioRiskPct:
          type: number
        withinLimits:
          type: boolean
        thetaDaily:
          type: number
        thetaMonthly:
          type: number
    PositionThreat:
      type: object
      required:
        [tradeId, symbol, symbolName, strategy, strategyLabel, threatLevel, severity, action, actionLabel, currentPnl, currentPnlPercent, currentPop, daysToExpiry, rationale]
      properties:
        tradeId:
          type: integer
        symbol:
          type: string
        symbolName:
          type: string
        strategy:
          type: string
        strategyLabel:
          type: string
        threatLevel:
          type: string
          enum: [green, yellow, red]
        severity:
          type: string
        action:
          type: string
        actionLabel:
          type: string
        currentPnl:
          type: number
        currentPnlPercent:
          type: number
        currentPop:
          type: number
        daysToExpiry:
          type: integer
        rationale:
          type: string
    ReportOpportunity:
      type: object
      required: [symbol, strategy, ravishScore, ravishTier, pop, ev, theta, eventRiskLevel]
      properties:
        symbol:
          type: string
        strategy:
          type: string
        ravishScore:
          type: integer
        ravishTier:
          type: string
        pop:
          type: number
        ev:
          type: number
        theta:
          type: number
        eventRiskLevel:
          type: string
    TradeToAvoid:
      type: object
      required: [symbol, symbolName, reason, level]
      properties:
        symbol:
          type: string
        symbolName:
          type: string
        reason:
          type: string
        level:
          type: string
        earningsInDays:
          type: integer
          nullable: true
    DailyReport:
      type: object
      required:
        [date, generatedAt, health, summary, market, positions, opportunities, tradesToAvoid, disclaimer]
      properties:
        id:
          type: integer
        date:
          type: string
        generatedAt:
          type: string
        briefing:
          type: string
        briefingSource:
          type: string
        health:
          $ref: "#/components/schemas/PortfolioHealth"
        summary:
          $ref: "#/components/schemas/ReportSummary"
        market:
          $ref: "#/components/schemas/MarketBriefing"
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionThreat"
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/ReportOpportunity"
        tradesToAvoid:
          type: array
          items:
            $ref: "#/components/schemas/TradeToAvoid"
        disclaimer:
          type: string
    CompareReportsRequest:
      type: object
      required: [fromId, toId]
      properties:
        fromId:
          type: integer
        toId:
          type: integer
    CompareReportsResponse:
      type: object
      required: [narration, narrationSource]
      properties:
        narration:
          type: string
        narrationSource:
          type: string
          enum: [llm, template]
    DailyReportSummary:
      type: object
      required:
        [id, reportDate, healthScore, healthLabel, exposureScore, riskScore, openPositions, redCount, yellowCount, netTheta, createdAt]
      properties:
        id:
          type: integer
        reportDate:
          type: string
        healthScore:
          type: integer
        healthLabel:
          type: string
        exposureScore:
          type: integer
        riskScore:
          type: integer
        openPositions:
          type: integer
        redCount:
          type: integer
        yellowCount:
          type: integer
        netTheta:
          type: number
        briefing:
          type: string
        createdAt:
          type: string
    DeleteReportResult:
      type: object
      required: [id, deleted]
      properties:
        id:
          type: integer
        deleted:
          type: boolean
    RestorableReport:
      type: object
      required:
        [reportDate, healthScore, healthLabel, exposureScore, riskScore, openPositions, redCount, yellowCount, netTheta, briefing, briefingSource, payload, createdAt]
      properties:
        reportDate:
          type: string
        healthScore:
          type: integer
        healthLabel:
          type: string
        exposureScore:
          type: integer
        riskScore:
          type: integer
        openPositions:
          type: integer
        redCount:
          type: integer
        yellowCount:
          type: integer
        netTheta:
          type: number
        briefing:
          type: string
        briefingSource:
          type: string
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
    ClearReportsResult:
      type: object
      required: [deleted, reports]
      properties:
        deleted:
          type: integer
        reports:
          type: array
          items:
            $ref: "#/components/schemas/RestorableReport"
    RestoreReportsRequest:
      type: object
      required: [reports]
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/RestorableReport"
    RestoreReportsResult:
      type: object
      required: [restored]
      properties:
        restored:
          type: integer